Implemented @Harmj0y changes
